navigation reacts on click at first time
